Add debug logs and add fix for searching dependencies for '*' variant within task group
